==> 1 patent/1_Extracting_vehicle_patent.ipynb <==
# Extracting vehicle patent

## 1. 사전 작업
### Open zipfile 
zip file read

import pandas as pd
import zipfile
import csv

zf_patent1 = zipfile.ZipFile('J:/data/patent/harvard_dataset_final_merged_output_1.zip') 
csv_patent1 = zf_patent1.open('harvard_dataset_final_merged_output_1.csv')

csv_patent_reader1 = csv.reader(csv_patent1)

### Where is info about class?
어떤 colummn이 class 정보를 담고 있는지 파악

일단 첫 번째 row의 value부터 살펴보자

header = csv_patent_reader1.next()
row1 = csv_patent_reader1.next()
for i in range(0,len(row1)):
    if row1[i] != '':
        print i, header[i], row1[i]

class 정보를 담고 있는 column이 몇 번째 column인지 살펴보자, 일단 62457부터 class 정보가 존재한다

for i in range(62457,len(header)+1):
    if 'class' in header[i]:
        print i, header[i]

그 결과, 62457 ~ 63353이 class 정보를 가진 column임을 알 수 있다

하지만 class가 숫자로 분류되어있다. 일단 10개정도의 row만 살펴보면서 어떤 숫자들이 있는지 알아보자

for i in range(1,11):
    row=csv_patent_reader1.next()
    for j in range(62457,63354):
        if 'class' in header[j] and row[j]!='':
            print header[j], row[j]

### 질문

1. 왜 patent 1개 (row 1개)에 mainclass, subclass가 최대 179개까지 존재하는가?
 - row 1개의 의미는? : patent 1개, class는 여러 개 가질 수 있음
 - main class만 기준으로 삼아도 되는가? : main class가 3digit level -> 분류하기 적당함
 - 숫자 외에 혹시 문자가 섞여있는 class가 있는가? :
 
2. class 숫자의 의미는?

https://www.uspto.gov/web/patents/classification/selectnumwithtitle.htm 
https://www.uspto.gov/web/patents/classification/international/ipc/ipc8/ipc_concordance/ipcsel.htm
자율주행 :  http://www.slideshare.net/LexInnova/driverless-car-technology-patent-landscape-analysis

44	Fuel and related compositions

70	Locks

73	Measuring and testing

74	Machine element or mechanism

123	Internal-combustion engines

152	Resilient tires and wheels

173	Tool driving or impacting

180	Motor vehicles

188	Brakes

191	Electricity: transmission to vehicles

192	clutches and power-stop control

248	Supports

278 Land vehicles: animal draft appliances

280	Land vehicles

293	Vehicle fenders

294	Handling: hand and hoist-line implements

296	Land vehicles: bodies and tops

297	Chairs and seats

298	Land vehicles: dumping

301	Land vehicles: wheels and axles

303	Fluid-pressure and analogous brake systems

305	Wheel substitutes for land vehicles

307	Electrical transmission or interconnection systems

310	Electrical generator or motor structure

315	Electric lamp and discharge devices: systems

318	Electricity: motive power systems

320	Electricity: battery or capacitor charging or discharging

322	Electricity: single generator systems

323	Electricity: power supply or regulation systems

324	Electricity: measuring and testing

326	Electronic digital logic circuitry

327	Miscellaneous active electrical nonlinear devices, circuits, and systems

340	Communications: electrical

341	Coded data generation or conversion

346	Recorders

361	Electricity: electrical systems and devices

363	Electric power conversion systems

368	Horology: time measuring systems or devices

370	Multiplex communications

375	Pulse or digital communications

377	Electrical pulse counters, pulse dividers, or shift registers: circuits and systems

388	Electricity: motor control systems

403	Joints and connections

429	Chemistry: electrical current producing apparatus, product, and process

439	Electrical connectors

474	Endless belt power transmission systems or components

476	Friction gear transmission systems or components

477	Interrelated power delivery controls, including engine control

492	Roll or roller

700	Data processing: generic control systems or specific applications

701	Data processing: vehicles, navigation, and relative location

702	Data processing: measuring, calibrating, or testing

706	Data processing: artificial intelligence

707	Data processing: database and file management or data structures

709	Electrical computers and digital processing systems: multicomputer data transferring

710	Electrical computers and digital data processing systems: input/output

711	Electrical computers and digital processing systems: memory

712	Electrical computers and digital processing systems: processing architectures and instruction processing (e.g., processors)

713	Electrical computers and digital processing systems: support

714	Error detection/correction and fault detection/recovery

718	Electrical computers and digital processing systems: virtual machine task or process management or task management/control

719	Electrical computers and digital processing systems: interprogram communication or interprocess communication (ipc)

726	Information security

901	Robots

903	Hybrid electric vehicles (hevs)

## 2. Extracting matched rows

### making class list
making list with the numbers of class column

import pandas as pd
import zipfile
import csv

zf_patent1 = zipfile.ZipFile('E:/patent/all_patent/harvard_dataset_final_merged_output_1.zip') 
csv_patent1 = zf_patent1.open('harvard_dataset_final_merged_output_1.csv')
csv_patent_reader1 = csv.reader(csv_patent1)

header = csv_patent_reader1.next()
mainclass_col = []

for j in range(62457,63354):
    if 'mainclass' in header[j]:   
        mainclass_col.append(j)

making csv file with vehicle class

classinfo = pd.read_csv('E:/patent/veh_class.csv')
veh_class = classinfo['class'].tolist()
veh_class = map(int,veh_class)

myfile1 = open("harvard_vehicle_patent1.csv", 'wb')
myfile_writer1 = csv.writer(myfile1,quoting=csv.QUOTE_ALL)
myfile_writer1.writerow(header)

for line in csv_patent_reader1:
    for i in mainclass_col:
        if line[i]=='':
                break
                
        else:
            try:            
                if int(line[i]) in veh_class:
                    myfile_writer1.writerow(line)
                    break
                
            except ValueError:
                pass                
        
       
myfile1.close()

zf_patent2 = zipfile.ZipFile('E:/patent/all_patent/harvard_dataset_final_merged_output_2.zip')
csv_patent2 = zf_patent2.open('harvard_dataset_final_merged_output_2.csv')
csv_patent_reader2 = csv.reader(csv_patent2)

csv_patent_reader2.next()
myfile2 = open("harvard_vehicle_patent2.csv", 'wb')
myfile_writer2 = csv.writer(myfile2,quoting=csv.QUOTE_ALL)
myfile_writer2.writerow(header)

for line in csv_patent_reader2:
    for i in mainclass_col:
        if line[i]=='':
                break
                
        else:
            try:            
                if int(line[i]) in veh_class:
                    myfile_writer2.writerow(line)
                    break
                
            except ValueError:
                pass                
        
       
myfile2.close()

How many cols does csv have?

zf_vh_ptnt1 = zipfile.ZipFile('E:/patent/harvard_vehicle_patent1.zip') 
vh_csv1 = zf_vh_ptnt1.open('harvard_vehicle_patent1.csv')
vh_csv_reader1 = csv.reader(vh_csv1)

header = vh_csv_reader1.next()

n_col=0
for line in vh_csv_reader1:
    n_col=n_col+1

print n_col
