==> 1 patent/2_patent_get_freq.ipynb <==
# Class별 freq 계산

import pandas as pd
import zipfile
import csv
import os

os.chdir("E:/patent")

zf_patent1 = zipfile.ZipFile('harvard_vehicle_patent1.zip') 
csv_patent1 = zf_patent1.open('harvard_vehicle_patent1.csv')
csv_patent_reader1 = csv.reader(csv_patent1)

header=csv_patent_reader1.next()
mainclass_col = []

for j in range(62457,63354):
    if 'mainclass' in header[j]:   
        mainclass_col.append(j)

freq = pd.read_csv('veh_class_freq.csv')
veh_class = freq['class'].tolist()
veh_class = map(int,veh_class)

counting 시작

for line in csv_patent_reader1:
    for i in mainclass_col:
        if line[i]=='':   
            break
    
        else:
            try:            
                if int(line[i]) in veh_class:
                # 목표를 포착했다 freq를 1 늘리자
                
                    for j in range(0,len(veh_class)):
                        if int(line[i])==freq.at[j,'class']:
                            freq.at[j,'freq'] = freq.at[j,'freq']+1
                            break
                            # 볼일 다 봤으니 탈출
                
            except ValueError:
                pass

zf_patent2 = zipfile.ZipFile('harvard_vehicle_patent2.zip') 
csv_patent2 = zf_patent2.open('harvard_vehicle_patent2.csv')
csv_patent_reader2 = csv.reader(csv_patent2)

for line in csv_patent_reader2:
    for i in mainclass_col:
        if line[i]=='':   
            break
    
        else:
            try:            
                if int(line[i]) in veh_class:
                # 목표를 포착했다 freq를 1 늘리자
                
                    for j in range(0,len(veh_class)):
                        if int(line[i])==freq.at[j,'class']:
                            freq.at[j,'freq'] = freq.at[j,'freq']+1
                            break
                            # 볼일 다 봤으니 탈출
                
            except ValueError:
                pass

freq.to_csv('veh_class_freq.csv',index=False)

# Assignee별 freq 계산

## 1. Where is info about class?
어떤 colummn이 assignee 정보를 담고 있는지 파악

import pandas as pd
import zipfile
import csv
import os

os.chdir("E:/patent")

zf_patent1 = zipfile.ZipFile('harvard_vehicle_patent1.zip') 
csv_patent1 = zf_patent1.open('harvard_vehicle_patent1.csv')
csv_patent_reader1 = csv.reader(csv_patent1)

header=csv_patent_reader1.next()
assignee_col = []

row 몇 개 value 보자

19495 부터 assignee col 시작

row1 = csv_patent_reader1.next()
for i in range(0,len(row1)):
    if row1[i] != '':
        print i, header[i], row1[i]

row1[19500]

row2 = csv_patent_reader1.next()
for i in range(19492,20000):
    if row2[i] != '':
        print i, header[i], row2[i]

row3 = csv_patent_reader1.next()
for i in range(19492,20000):
    if row3[i] != '':
        print i, header[i], row3[i]

## 2. Data Problems
### 1. assignee ID를 알아볼 수 없음
organization 사용

### 2. assignee ID = organization?  
yes

### 3. assignee0 = assingee1 & organization0 = organization1?:  
mostly, but not necessarily  
  
### 4. difference between assignee0 & 1  
rawlocation, type  
rawlocation : 중복 시 "" or 잘못된 정보(대부분)  
type : 정체 불명, 중복 시 rawlocation=""이면 "", rawlocation!=""이면 숫자  
__=> raw data 문제!! 일단 넘긴다__

# prob 2
ass_list=[]
org_list=[]
for i in range(0,200000):
    line=csv_patent_reader1.next()
    ass_list.append(line[19495])
    org_list.append(line[19500])

df=pd.DataFrame({'assignee':ass_list, 'organization':org_list})
df=df.drop_duplicates()
dups=df.duplicated(subset='organization')
sum(dups)

# prob 3

check_ass=0
check_org=0
check_same=0
for i in range(0,1000):
    line=csv_patent_reader1.next()
    if line[19495]==line[19506]:
        check_ass=check_ass+1
    if line[19500]==line[19511]:
        check_org=check_org+1
    if line[19495]==line[19506] and line[19500]==line[19511]:
        check_same=check_same+1
    if line[19495]!=line[19506] and line[19500]!=line[19511]:
        print line[19495]+' / '+line[19506]+' / '+line[19500]+' / '+line[19511]

print check_ass, check_org, check_same

## 3. location 무시하고 assignee list 제작

#basic settings
import pandas as pd
import zipfile
import csv
import os

os.chdir("E:/patent")

csv_patent1 = open('harvard_vehicle_patent1.csv')
csv_patent_reader1 = csv.reader(csv_patent1)
header=csv_patent_reader1.next()

zf_patent2 = zipfile.ZipFile('harvard_vehicle_patent2.zip') 
csv_patent2 = zf_patent2.open('harvard_vehicle_patent2.csv')
csv_patent_reader2 = csv.reader(csv_patent2)
csv_patent_reader2.next()

#select the assignee col
assignee_col = []
for i in range(0,len(header)):
    if 'organization' in header[i]:   
        assignee_col.append(i)
        
del assignee_col[-3:]    #delete lawyer org

#return assignees
assignee = []
p_id = []
row = []
for line in csv_patent_reader1:
    for i in assignee_col:
        if line[i]=='':
            break
        
        elif line[i]=='nan':
            assignee.append('private')
            p_id.append(line[1])
            row.append(line[0])
            
        else:
            assignee.append(line[i])
            p_id.append(line[1])
            row.append(line[0])
            
for line in csv_patent_reader2:
    for i in assignee_col:
        if line[i]=='':
            break
        
        elif line[i]=='nan':
            assignee.append('private')
            p_id.append(line[1])
            row.append(line[0])
            
        else:
            assignee.append(line[i])
            p_id.append(line[1])
            row.append(line[0])
            
df = pd.DataFrame({'patent_id':p_id,'assignee':assignee,'row':row})
df = df.drop_duplicates()
df.to_csv('veh_pid_assignee.csv',index=False)

## 4. Assignee 이름 정리해서 다시 제작
### 필요한 이유  
assignee - company, company name 기준으로 matching만 완료  
company name이 없는 assignee = 투자 받지 않은 assignee는? : 엉망진창  
따라서 나중에 patent 단위로 control group을 보기 위해서 patent assignee의 이름을 정리

import pandas as pd
import csv
import os

os.chdir("E:/patent")
df=pd.read_csv('veh_pid_assignee.csv')
df['new_assignee']=''

def nameprocessing(s):
    s = str(s)
    
    s = s.replace(".","")
    s = s.replace(",","")    
    s = s.replace(";","")
    s = ''.join(s.split()).lower()
    
    s = s.replace("incorporated","inc")
    s = s.replace("limited","ltd")
    s = s.replace("gmbh","")
    
    rm=['inc','ltd','llc','pte']
    for i in range(0,len(s)):
        if s[-3:] in rm:
            s=s[:-3]
        else:
            break

    return s

for i in range(0,len(df.index)):
    df.iloc[i,3]=nameprocessing(df.iloc[i,0])
    
df.to_csv('veh_pid_new_assignee.csv',index=False)
freq=df.groupby(['new_assignee']).size().reset_index().rename(columns={'0':'count'})
freq.columns = ['new_assignee','freq']
freq.to_csv('veh_new_assignee_freq.csv',index=False)

너무 오래 걸려서 interrupt, checkpoint 확인

print df.iloc[1202130,3], df.iloc[1202131,3]

그래도 거의 다 했네... checkpoint = 1202130까지 완료  
저장해놓고 다음에 하자

df.to_csv('veh_pid_assignee.csv',index=False)

import pandas as pd
import csv
import os

os.chdir("E:/patent")
df=pd.read_csv('veh_pid_assignee.csv')

for i in range(1202130,len(df.index)):
    df.iloc[i,3]=nameprocessing(df.iloc[i,0])
    
df.to_csv('veh_pid_assignee.csv',index=False)
freq=df.groupby(['new_assignee']).size().reset_index().rename(columns={'0':'count'})
freq.columns = ['new_assignee','freq']
freq.to_csv('veh_new_assignee_freq.csv',indeax=False)
