==> 1 patent/3_get selected patent info.ipynb <==
# raw data에서 Patent 변수 생성에 필요한 info만 걸러내기

### raw data(patent)에서 필요한 정보

1. patent id (이미 완료)
2. assignee (이미 완료)
3. patent date (나중에 year로 변환)
4. patent class
5. citation edge list

### 1. patent year / class

__patent granted year 정보가 없음...=> 교수님과 이야기 필요__  
    대안: citation에 date 정보 (year, month) 있음  
    문제: citation 0인 patent는 date 확인 불가  
    일단 class만 만들고 citation 만들 때 year까지 집어넣자 

__교수님과 이야기 완료__   
class: __subclass까지 다 뽑기__   
date: __raw data 새로 받아서__ 뽑기

### 1.1 main class 뽑기

import pandas as pd
import zipfile
import csv
import os

os.chdir("E:/patent")

zf_patent1 = zipfile.ZipFile('harvard_vehicle_patent1.zip') 
csv_patent1 = open('harvard_vehicle_patent1.csv')
csv_patent_reader1 = csv.reader(csv_patent1)
header=csv_patent_reader1.next()

zf_patent2 = zipfile.ZipFile('harvard_vehicle_patent2.zip') 
csv_patent2 = zf_patent2.open('harvard_vehicle_patent2.csv')
csv_patent_reader2 = csv.reader(csv_patent2)
csv_patent_reader2.next()

mainclass_col = []
for j in range(62457,63354):
    if 'mainclass' in header[j]:   
        mainclass_col.append(j)

with open('veh_pid_class.csv','wb') as myfile:
    myfile_writer = csv.writer(myfile,quoting=csv.QUOTE_ALL)
    
    for line in csv_patent_reader1:
        row = []
        row.append(line[0])
        row.append(line[1])
        for i in mainclass_col:
            if line[i] == '':
                break
            else:
                row.append(line[i])
        myfile_writer.writerow(row)
        
    for line in csv_patent_reader2:
        row = []
        row.append(line[0])
        row.append(line[1])
        for i in mainclass_col:
            if line[i] == '':
                break
            else:
                row.append(line[i])
        myfile_writer.writerow(row)

__문제 발생__  
class에서 중복 발견  
=> line by line 중복 처리 필요 

csv_patent1 = open('veh_pid_class.csv','rb')
csv_patent_reader1 = csv.reader(csv_patent1)
header=csv_patent_reader1.next()

with open('veh_pid_class_no_dup.csv','wb') as myfile:
    myfile_writer = csv.writer(myfile,quoting=csv.QUOTE_ALL)
    myfile_writer.writerow(header)
    for line in csv_patent_reader1:
        no_dup = list(set(line[2:]))
        row = line[0:2] + no_dup
        myfile_writer.writerow(row)

df=pd.read_csv('veh_pid_class_no_dup.csv')

column 몇 개로 해야하는지 확인좀 (현재 20개 이하인건 확실)

check = df.iloc[:,20:22]
check.iloc[0:3,]

len(check.index)*2

check.isnull().sum().sum()

오오 19에서 null 아닌 친구가 있나봄, 확인 ㄱㄱ

cl19 = check.iloc[:,0]
cl20 = check.iloc[:,1]
print len(cl19), cl19.isnull().sum(), cl20.isnull().sum()

good~ class 개수는 __19개__가 적당하군

### 1.2 main, sub 다 뽑기

import pandas as pd
import zipfile
import csv
import os

os.chdir("E:/patent")

zf_patent1 = zipfile.ZipFile('harvard_vehicle_patent1.zip') 
csv_patent1 = open('harvard_vehicle_patent1.csv')
csv_patent_reader1 = csv.reader(csv_patent1)
header=csv_patent_reader1.next()

zf_patent2 = zipfile.ZipFile('harvard_vehicle_patent2.zip') 
csv_patent2 = zf_patent2.open('harvard_vehicle_patent2.csv')
csv_patent_reader2 = csv.reader(csv_patent2)
header=csv_patent_reader2.next()

class_col = []
for j in range(62457,63354):
    if 'class' in header[j]:   
        class_col.append(j)

head

head = ['row','patent_id']
head = head + list(header[i] for i in class_col)

with open('veh_pid_class.csv','wb') as myfile:
    myfile_writer = csv.writer(myfile,quoting=csv.QUOTE_ALL)
    myfile_writer.writerow(head)
    
    for line in csv_patent_reader1:
        row = []
        row.append(line[0])
        row.append(line[1])
        for i in class_col:
            if line[i] == '':
                break
            elif 'class' in line[i]:
                break
            else:
                row.append(line[i])
        myfile_writer.writerow(row)
        
    for line in csv_patent_reader2:
        row = []
        row.append(line[0])
        row.append(line[1])
        for i in class_col:
            if line[i] == '':
                break
            else:
                row.append(line[i])
        myfile_writer.writerow(row)

with open('veh_pid_class.csv','rb') as csvfile:
    csv_reader = csv.reader(csvfile)
    maxlen = 0
    for line in csv_reader:
        a=len(line)
        if a > maxlen:
            maxlen = a
            
maxlen

subclass에 mainclass 정보까지 불필요하게 있음  
__=> mainclass 부분 제거 후 int로 만들자__

csvin = open('veh_pid_class.csv','rb')
csv_reader = csv.reader(csvin)
header = csv_reader.next()

subclass_col = []
for i in range(0,len(header)):
    if 'subclass' in header[i]:   
        subclass_col.append(i)

def to_int(row):    
    try:
        row=map(int,row)
    except ValueError:
        try:
            for i in range(0,len(row)):
                row[i]=int(row[i])
        except ValueError:
            pass  
        
    return row

with open('veh_pid_class_processed.csv','wb') as csvout:    
    csv_writer = csv.writer(csvout,quoting=csv.QUOTE_ALL)
    csv_writer.writerow(header)
    for line in csv_reader:
        if int(line[0]) > 2964076 :
            for i in subclass_col:
                try:
                    if line[i]=='':
                        break
                    else:
                        line[i]=line[i].split('/')[1]
                except IndexError:
                    break
            csv_writer.writerow(to_int(line))
            
        else:
            csv_writer.writerow(to_int(line))

### 2. citation edge list + date

patent_id, citation_id, date 해서 쭈루루룩 나열

import pandas as pd
import zipfile
import csv
import os

os.chdir("E:/patent")

zf_patent1 = zipfile.ZipFile('harvard_vehicle_patent1.zip') 
csv_patent1 = open('harvard_vehicle_patent1.csv')
csv_patent_reader1 = csv.reader(csv_patent1)
header=csv_patent_reader1.next()

zf_patent2 = zipfile.ZipFile('harvard_vehicle_patent2.zip') 
csv_patent2 = zf_patent2.open('harvard_vehicle_patent2.csv')
csv_patent_reader2 = csv.reader(csv_patent2)
header1=csv_patent_reader2.next()

citation_col = []
for j in range(20377,62457):
    if 'citation' in header[j]:   
        citation_col.append(j)

with open('veh_pid_citation_duplicated.csv','wb') as myfile:
    myfile_writer = csv.writer(myfile,quoting=csv.QUOTE_ALL)    
    for line in csv_patent_reader1:
        for i in citation_col:
            if line[i] == '':
                break
            else:
                try:
                    row = []
                    row.append(int(line[1]))
                    row.append(int(line[i]))
                    row.append(line[i+1])
                    myfile_writer.writerow(row)
                
                except ValueError:
                    pass
        
    for line in csv_patent_reader2:
        for i in citation_col:
            if line[i] == '':
                break
            else:
                try:
                    row = []
                    row.append(int(line[1]))
                    row.append(int(line[i]))
                    row.append(line[i+1])
                    myfile_writer.writerow(row)
                
                except ValueError:
                    pass

#remove duplicates
df=pd.read_csv('veh_pid_citation_duplicated.csv')
df=df.drop_duplicates()
df.to_csv('veh_pid_citation_no_dup.csv',index=False)

### 2.1 citation edge list 만들기

df=pd.read_csv('veh_pid_citation_no_dup.csv')
df = df.iloc[:,0:2]
list(df)

len(df)

df.drop_duplicates(inplace=True)
len(df)

df.to_csv('veh_citation_network.csv',index=False)

### 2.2 patent date 뽑기

df=pd.read_csv('veh_pid_citation_no_dup.csv')
list(df)

df = df.iloc[:,1:3]
df.rename(columns={'cited_id': 'patent_id'}, inplace=True)
list(df)

pid = pd.read_csv('veh_pid_class_no_dup.csv')
pid = pid.iloc[:,1]
pid[0:3,]

pid = pid.to_frame()
type(pid)

pid_date = pid.merge(df,how='left',on='patent_id',sort=True)
pid_date.iloc[0:30,]

pid_date['granted_date'].replace({'0000-00-00': np.nan}, inplace=True)
pid_date.iloc[0:30,]

pid_date = pid_date.drop_duplicates()
pid_date.dropna(axis=0, inplace=True)
print len(pid_date), len(pid)

final = pid.merge(pid_date,how='left',on='patent_id',sort=True)
print len(final),len(pid)

final.isnull().sum()

1. __date가 없는 patent__가 45만개
2. __같은 patent_id, 다른 granted_date__가 7000개

for i in range(0,len(final)):
    if final.iloc[i,0] == final.iloc[i+1,0]:
        print i
        break

final.iloc[1484:1488,]

모르겠다 일단 저장해놓자

final.to_csv('veh_pid_date.csv',index=False)

### 2.3 새 raw data 에서 date 뽑기

import os
import pandas as pd

os.chdir("E:/patent")
df = pd.read_csv('all_patent/grant_application.csv')
list(df)

df = df.iloc[:,(1,5)]
list(df)

__patent_id가 이상한 애들 처리 필요!__

__merge ㄱㄱ__

pid = pd.read_csv('veh_pid_date.csv')
pid = pid.iloc[:,0]
pid = pid.to_frame()

pid_date = pid.merge(df,how='left',on='patent_id',sort=True)
pid_date.iloc[0:10,]

print len(pid), len(pid_date)

pid_date.to_csv('veh_pid_date.csv', index=False)
